node 1 works

+++ Node2/can.h
@@ -71,7 +71,7 @@ struct Byte8 {
 //    // Should print: CanMsg(id:1, length:7, data:{10, 0, 20, 0, 0, 240, 193})
 typedef struct CanMsg CanMsg;
 struct CanMsg {
-    uint8_t id;
+    uint16_t id;
     uint8_t length;
     union {
         uint8_t     byte[8];
